utils: move non-vscode-dependent utils to utils/
Refactor use of killProcessTree to a single function in utils/processUtils

GitHub-Last-Rev: 29d0561c3412d0d4f470bc9e516bca7303f7cf4f
GitHub-Pull-Request: golang/vscode-go#344
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/242518
