+++ src/goMain.ts
@@ -35,7 +35,6 @@ import { lintCode } from './goLint';
 import { GO_MODE } from './goMode';
 import { addTags, removeTags } from './goModifytags';
 import { GO111MODULE, isModSupported } from './goModules';
-import { clearCacheForTools, envPath, fileExists, getCurrentGoRoot, setCurrentGoRoot } from './goPath';
 import { playgroundCommand } from './goPlayground';
 import { GoReferencesCodeLensProvider } from './goReferencesCodelens';
 import { GoRunTestCodeLensProvider } from './goRunTestCodelens';
@@ -60,6 +59,7 @@ import {
 	handleDiagnosticErrors,
 	isGoPathSet
 } from './util';
+import { clearCacheForTools, envPath, fileExists, getCurrentGoRoot, setCurrentGoRoot } from './utils/goPath';
 
 export let buildDiagnosticCollection: vscode.DiagnosticCollection;
 export let lintDiagnosticCollection: vscode.DiagnosticCollection;
@@ -73,7 +73,8 @@ export let restartLanguageServer = () => { return; };
 export async function activate(ctx: vscode.ExtensionContext): Promise<void> {
 	if (!getBinPath('go', false)) {
 		const choice = await vscode.window.showInformationMessage(
-			`Go could not be found in either GOROOT(${getCurrentGoRoot()}) or PATH(${envPath}. Install Go and reload the window.`,
+			`Go could not be found in either GOROOT(${getCurrentGoRoot()}) or PATH(${envPath}. ` +
+			`Install Go and reload the window.`,
 			'Go to Download Page'
 		);
 		if (choice === 'Go to Download Page') {
