+++ src/goOutline.ts
@@ -13,9 +13,9 @@ import {
 	getBinPath,
 	getFileArchive,
 	getGoConfig,
-	killProcess,
 	makeMemoizedByteOffsetConverter
 } from './util';
+import {killProcess} from './utils/processUtils';
 
 // Keep in sync with https://github.com/ramya-rao-a/go-outline
 export interface GoOutlineRange {
